Add in room management (in URL path)
With the -rooms options, gosses will now expect a "room" just after
the prefix part of the URL, i.e. something like:

http://example.com/my-prefix/myroom/rest-of-path

In this mode, listing of http://example.com/my-prefix is forbidden. The
room name must be known in advance. This feature offers some kind of
containerization of the served content, making it possible to share an
URL for a subset of the served content.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>